backend/pyproject.toml
[project]
name = "meal-agent"
version = "0.1.0"
description = "Meal-decision agent backend — generic, brand-agnostic"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.32",
  "pydantic>=2.9",
  "pydantic-settings>=2.6",
  "langgraph>=0.2.50",
  "langgraph-checkpoint-postgres>=2.0",
  "langchain>=0.3",
  "langchain-openai>=0.2",
  "langchain-mcp-adapters>=0.1",
  "mcp>=1.0",
  "asyncpg>=0.30",
  "psycopg[binary,pool]>=3.2",
  "pyyaml>=6.0",
  "structlog>=24",
  "httpx>=0.27",
  "aiohttp>=3.13.5",
]

[project.optional-dependencies]
dev = [
  "pytest>=8",
  "pytest-asyncio>=0.24",
  "respx>=0.21",
  "ruff>=0.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/meal_agent"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "ASYNC"]
